+++ modules/function-app-flex/examples/basic/outputs.tf
@@ -1,0 +1,11 @@
+output "function_app_id" {
+  value = module.function_app_flex.id
+}
+
+output "function_app_name" {
+  value = module.function_app_flex.name
+}
+
+output "default_hostname" {
+  value = module.function_app_flex.default_hostname
+}